improved download script + added some more data

+++ Scripts/meteo.ps1
@@ -1,7 +1,7 @@
 . "$PSScriptRoot\gzip.ps1"
 
 Function Get-Meteo-File {
-    Param([Int]$year, [Int]$month, [String]$destDir)
+    Param([Int]$year, [Int]$month, [String]$destDir, [Boolean] $overWrite = ($false))
     $csvFileName = "synop." + $year.ToString("0000") + $month.ToString("00") + ".csv"
     $gzFileName = $csvFileName + ".gz"
     $urlBase = "https://donneespubliques.meteofrance.fr/donnees_libres/Txt/Synop/Archive/"
@@ -9,22 +9,29 @@ Function Get-Meteo-File {
     $gzFilePath = $destDir + $gzFileName
     $csvFilePath = $destDir + $csvFileName
 
+    if ((-not $overWrite) -and (Test-Path($csvFilePath))) {
+        return
+    }
+
     Write-Output ("Downloading " + $url)
-    try { Invoke-WebRequest -Uri $url -OutFile $gzFilePath }
+    try { 
+        Invoke-WebRequest -Uri $url -OutFile $gzFilePath 
+        Start-Sleep -Milliseconds 500 # don't DDoS Meteo France's servers
+    }
     catch { 
         Write-Output "Error during download :"
         Write-Output $_
         try { Remove-Item $gzFilePath } catch { }
-        continue 
+        return 
     }
-    
+
     Write-Output ("decompressing " + $gzFileName)
     try { Expand-GZip-File $gzFilePath $csvFilePath }
     catch { 
         Write-Output "Error during decompression :"
         Write-Output $_
         try { Remove-Item $csvFilePath } catch { }
-        continue 
+        return 
     }
     finally { Remove-Item $gzFilePath }
 }
